@@ -6,7 +6,22 @@ function ProjectList({ projects }) {
   return (
     <div id="projects">
       <h2>My Projects</h2>
-      <div id="project-list">{/* render ProjectItem components here */}</div>
+      <div id="project-list">
+         {projects.map((project)=> (
+          <ProjectItem
+           key = {project.id}
+           id={project.id}
+           name={project.name}
+           about={project.about}
+           technologies={project.technologies}
+          />
+        )
+
+        )
+        }
+
+
+      </div>
     </div>
   );
 }
